web/requirements.txt
Flask
pymysql
python-dotenv
bcrypt